Download of hourly data from automatic weather stations (AWS) of INMET-Brazil

This function will download the hourly AWS data of INMET for whatever station of interest, based on the period of time selected (start_date and end_date) and station code. The limit acquisition in the same requisition for hourly data is one year.

hourly_weather_station_download(start_date, end_date, station)

Arguments

start_date

Date that start the investigation, should be in the following format (1958-01-01 /Year-Month-Day)

end_date

Date that end the investigation, should be in the following format (2017-12-31 /Year-Month-Day)

station

The station code (ID - WMO code) for download. To see the station ID, please see the function *see_stations_info*.

Value

Returns a data.frame with the AWS data requested

Author

Roberto Filgueiras

Examples

if (FALSE) { # \dontrun{
df <- hourly_weather_station_download(
                                      station = "A001",
                                      start_date = "2022-08-12",
                                      end_date = "2022-08-16")
} # }